@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 0.1.5
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
