@@ -4,13 +4,13 @@ version: 0.2.4+1
 homepage: https://github.com/smaho-engineering/esptouch_flutter
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: '>0.9.0 <2.0.0'
+  meta:
 
 flutter:
   plugin:
